requirements_desktop.txt
fastapi==0.104.1
uvicorn==0.24.0
pywebview==4.4.1
pyinstaller==6.2.0
